lambda_build/backend/requirements.txt
fastapi>=0.110.0
mangum>=0.17.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
mistralai>=1.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
